pyproject.toml
[project]
name = "nanachat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "regex>=2025.10.23",
    "requests>=2.32.5",
]
